requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==0.12

# Flask Extensions
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-User==0.6.11
Flask-WTF==0.13.1

# Python packages
hashids==1.2.0
Unidecode==0.4.20

# Development tools
pytest==3.0.5
pytest-flask==0.10.0
pytest-cov==2.4.0

